+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Add support for SRDF descriptions (thanks to @nickswalker)
+- Description: add SRDFs for Franka Emika robots: Panda, FER, FR3, FR3v2, FR3v2_1 (thanks to @nickswalker)
+- Description: franka_description repository updated for SRDF Xacro updates
+
+
 ## [2.0.0] - 2026-05-05
 
 ### Added